continous progress bar

+++ project/js/pot.js
@@ -12,11 +12,11 @@ function display() {
     $('#time').text(seconds + '.' + millisec);
     //fix fullstack error, use literals with care...
     timer = setTimeout('display()', 100);
-    var totalTime = parseInt(localStorage.getItem('timePerTurn'));
+    var totalTime = parseFloat(localStorage.getItem('timePerTurn'));
     if (seconds == totalTime) {
         $('#time').trigger('timeout');
     }
-    var percentage = parseInt(seconds + '.' + millisec) / totalTime * 100;
+    var percentage = parseFloat(seconds + '.' + millisec) / totalTime * 100;
     $('.progress-bar')
         .attr('style', 'width: ' + percentage + '%')
         .attr('aria-valuenow', seconds + '.' + millisec);
